refs(subscriptions): Start using fields from `snuba_query` to replace fields on the alert rule
This changes us to use fields on the `snuba_query` in more places. I'll do another pass at this
after this pr and catch the rest that I missed. It's fine to read from these fields in either the
snuba_query or alert_rule, since we're keeping the two in sync at the moment.